[RSC-112] Introduce blocked update to bank/user
This commit introduces the query of blocks from bank in fixed sized
blocks. Current size is max 50 blocks or less (delta / 10) and is
hardcoded in user Operations.hs. Also this commit removes bank's
GetHBlock endpoint, but all Communication methods were altered to
support the new design and should not fail. This also means that the
old client versions wouldn't work with new bank as update request
format is different.