[RSC-33] Remove comments about closing state in tests.
I was thinking about it and decided that it is not necessary at all.
Closing state is intended to release some resources. And maybe to
improve persistence. But we don't need any persistence in tests and the only
resource acquired is memory, which is not resource in haskell.
Another reason is that in memory mode `closeAcidState` only prevents state from
being used after close. That's all.